+++ README.md
@@ -5,3 +5,53 @@
 ## 우아한테크코스 코드리뷰
 
 - [온라인 코드 리뷰 과정](https://github.com/woowacourse/woowacourse-docs/blob/master/maincourse/README.md
+## 실행 프로세스
+
+블랙잭 게임을 변형한 프로그램을 구현한다. 블랙잭 게임은 딜러와 플레이어 중 카드의 합이 21 또는 21에 가장 가까운 숫자를 가지는 쪽이 이기는 게임이다.
+
+카드의 숫자 계산은 카드 숫자를 기본으로 하며, 예외로 Ace는 1 또는 11로 계산할 수 있으며, King, Queen, Jack은 각각 10으로 계산한다.
+게임을 시작하면 플레이어는 두 장의 카드를 지급 받으며, 두 장의 카드 숫자를 합쳐 21을 초과하지 않으면서 21에 가깝게 만들면 이긴다. 21을 넘지 않을 경우 원한다면 얼마든지 카드를 계속 뽑을 수 있다.
+딜러는 처음에 받은 2장의 합계가 16이하이면 반드시 1장의 카드를 추가로 받아야 하고, 17점 이상이면 추가로 받을 수 없다.
+게임을 완료한 후 각 플레이어별로 승패를 출력한다.
+
+1. 게임에 참여할 사람의 이름 입력
+2. 딜러와 플레이어에게 2장씩 카드를 나눠줌
+3. 각 플레이어에게 카드 더 받을건지 물어봄
+   3-1. 플레이어가 "n"를 입력할 때까지 반복
+4. 딜러의 카드가 17 이상이 될 때까지 카드를 더 받음
+5. 딜러와 플레이어 카드 출력
+6. 최종 승패 출력
+
+## 기능목록
+
+
+- 참가자 (플레이어 + 딜러)
+    - [x] 카드를 받아서 가지고 있음
+    - [x] 모든 카드의 숫자 합을 반환하는 기능
+- 플레이어
+    - [x] validation: 플레이어 이름은 중복 불가
+    - [x] validation: 플레이어 이름의 길이 1글자 이상 10글자 이하
+    - [x] 쉼표 기준으로 분리
+    - [x] validation: 플레이어의 수는 최소 1명 최대 7명
+    - [x] 카드 2장을 받음
+    - [x] 매 턴마다 카드 숫자가 21 초과 인지 확인
+    - [x] 카드를 받음
+- 딜러
+    - [x] 카드 2장을 받음
+    - [x] 매 턴마다 카드 숫자가 16 초과 인지 확인
+    - [x] 카드를 받음
+- 게임
+    - [x] 참가자들에게 초기 카드 배분
+    - [x] 참가자들에게 각 턴마다 카드 배분
+    - [x] 딜러와 플레이어의 결과를 비교하여 승패 판단
+- 카드
+    - [x] 카드 숫자 반환
+  - 패
+    - [x] 
+- 덱
+    - [x] 모든 카드를 가짐
+
+- ACE
+    - [x] 모든 카드의 합이
+        - [x] 21보다 작으면 11로 적용
+        - [x] 21보다 크면 1로 적용
