+++ src/main/java/domain/card/ShuffleStrategy.java
@@ -1,6 +1,8 @@
-package domain;
+package domain.card;
 
 import java.util.Stack;
+
+import domain.card.Card;
 
 public interface ShuffleStrategy {
     Stack<Card> shuffle(Stack<Card> deck);
