+++ src/main/java/domain/game/PrintedHandPool.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.game;
 
 import java.util.HashSet;
 import java.util.List;
