+++ src/main/java/service/GameService.java
@@ -5,12 +5,12 @@ import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
 
-import domain.Deck;
-import domain.Participant;
-import domain.Participants;
-import domain.Player;
-import domain.PrintedHandPool;
-import domain.Result;
+import domain.card.Deck;
+import domain.people.Participant;
+import domain.people.Participants;
+import domain.people.Player;
+import domain.game.PrintedHandPool;
+import domain.game.Result;
 
 public class GameService {
     private static final int DEALER_MINIMUM_VALUE= 17;
