+++ src/test/java/domain/ParticipantTest.java
@@ -9,6 +9,12 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import domain.card.Card;
+import domain.card.Rank;
+import domain.card.Suit;
+import domain.people.Dealer;
+import domain.people.Participant;
+
 class ParticipantTest {
     private Participant participant;
