+++ src/test/java/domain/PlayerTest.java
@@ -7,6 +7,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 
+import domain.people.Player;
+
 class PlayerTest {
     @ParameterizedTest(name = "플레이어의 이름은 null이거나 공백일 수 없다.")
     @NullAndEmptySource
